requirements.txt
requests~=2.31.0
python-dotenv~=1.0.1
aiogram~=3.3.0
beautifulsoup4~=4.12.3
APScheduler~=3.10.4
pillow~=10.2.0
aioredis~=2.0.1